refactor(core): move groupBy, firstString, stringArray from memory-improve.ts to common.ts
Fixes #334
