+++ src/core/common.ts
@@ -369,3 +369,41 @@ function parseRetryAfter(response: Response): number | undefined {
 export function toErrorMessage(error: unknown): string {
   return error instanceof Error ? error.message : String(error);
 }
+
+// ── Generic data utilities ───────────────────────────────────────────────────
+
+/**
+ * Return the trimmed string if non-empty, otherwise `undefined`.
+ * Equivalent to `firstString` previously defined in `memory-improve.ts`.
+ */
+export function firstString(value: unknown): string | undefined {
+  return typeof value === "string" && value.trim().length > 0 ? value.trim() : undefined;
+}
+
+/**
+ * Coerce an unknown value to a filtered, trimmed string array.
+ * Non-strings and empty/whitespace-only entries are dropped.
+ */
+export function stringArray(value: unknown): string[] {
+  if (!Array.isArray(value)) return [];
+  const out: string[] = [];
+  for (const item of value) {
+    if (typeof item === "string" && item.trim().length > 0) out.push(item.trim());
+  }
+  return out;
+}
+
+/**
+ * Group an array of values by a string key derived from each element.
+ * Returns a `Map` so insertion order within each group is preserved.
+ */
+export function groupBy<T>(values: T[], keyFn: (value: T) => string): Map<string, T[]> {
+  const groups = new Map<string, T[]>();
+  for (const value of values) {
+    const key = keyFn(value);
+    const existing = groups.get(key);
+    if (existing) existing.push(value);
+    else groups.set(key, [value]);
+  }
+  return groups;
+}
